+++ hurocon/meta.py
@@ -1,2 +1,3 @@
+name = 'hurocon'
 version = '0.5.0dev0'
 author = 'maximilionus'